fix: update files from imageService

+++ Backend/product_domain/imageService/deleteImage/src/controllers/deleteController.js
@@ -1,13 +1,13 @@
 const Image = require('../models/Image');
-// const { checkProductExists } = require('../services/productService');
+const { checkProductExists } = require('../services/productService');
 
 const deleteImagesByProductId = async (req, res) => {
   const { product_id } = req.params;
 
-  // const exists = await checkProductExists(product_id);
-  // if (!exists) {
-  //   return res.status(404).json({ error: 'Product not found in catalogService' });
-  // }
+  const exists = await checkProductExists(product_id);
+  if (!exists) {
+    return res.status(404).json({ error: 'Product not found in catalogService' });
+  }
 
   try {
     await Image.deleteMany({ product_id });
